moved files to src folder
split into multiple files
