+++ src/examples/template.tex
